+++ changelog.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `getTagData()` and `getFunctionData()` lucee compats
+
 ## [1.10.0] - 2024-10-28
 
 ### Fixed